The type of the field should always be preferred
All methods of a class are considered getter methods in gwt-jackson, the only condition is that they have a single parameter. This is of course ridiculous. 

Then, when getter visibility is NONE and only the field should be used, the type is still taken from the nonsense getter method that was detected. This hack switches to the field type first if there is a field, tests pass.